Moves filters to its own package and provides initial tests for composing functionality
